added context api for sharing data, refactored codes, removed unused packages
