+++ src/App.js
@@ -1,11 +1,7 @@
 import PaleoTable from './table/PaleoTable';
 
 function App() {
-  return (
-    <div className="App">
-      <PaleoTable width="1600px" />
-    </div>
-  );
+  return <PaleoTable />;
 }
 
 export default App;
